feat(worker): add worker to deliver webhooks

+++ db/entities/endpoint.go
@@ -82,5 +82,5 @@ func (m *Retry) Validate() error {
 }
 
 type FixedStrategyConfig struct {
-	Attempts []int `json:"attempts" default:"[0,60,3600]"`
+	Attempts []int64 `json:"attempts" default:"[0,60,3600]"`
 }
